feat: improve container classes and build system
* Enhance container class hierarchy
  - Add MulleObjCContainer base class
  - Move common container code to base class
  - Rename NSEnumerator+NSArray to NSArray+NSEnumerator
  - Add MulleObjCContainer+NSArray category

* Improve build system and CMake support
  - Add CMake package config files
  - Update compiler flags and warnings
  - Add TAO debug support via OBJC_TAO_DEBUG_ENABLED option
  - Fix Windows compatibility issues
  - Add sccache support for Linux

* Refactor source organization
  - Move shared code to MulleObjCContainer
  - Update header includes and dependencies
  - Fix file paths and directory structure
  - Clean up obsolete environment settings

+++ src/NSArray/_MulleObjCConcreteArray.m
@@ -263,7 +263,7 @@ static NSUInteger   findObjectWithRange( _MulleObjCConcreteArray *self,
 {
    if( ! state->state)
    {
-      state->mutationsPtr = (unsigned long *)self;
+      state->mutationsPtr = (NSUInteger *) self;
       state->itemsPtr     = _objects;
       state->state        = 1;
       return( self->_count);
